3.4.5
- Improved the interface for the virtual hard drive creation dialog box.
- Removed import HDD1.
