+++ app/proguard-rules.pro
@@ -24,8 +24,8 @@
 -keep class com.vectras.qemu.MainSettingsManager$UserInterfacePreferencesFragment { *; }
 -keep class com.vectras.qemu.MainSettingsManager$QemuPreferencesFragment { *; }
 -keep class com.vectras.qemu.MainSettingsManager$VncPreferencesFragment { *; }
--keep class com.vectras.vm.RequestNetwork { *; }
--keep class com.vectras.vm.RequestNetworkController { *; }
+-keep class com.vectras.vm.network.RequestNetwork { *; }
+-keep class com.vectras.vm.network.RequestNetworkController { *; }
 -keep class android.media.LoudnessCodecController { *; }
 -keepclassmembers class com.google.firebase.database.GenericTypeIndicator{*;}
 -keep class * extends com.google.firebase.database.GenericTypeIndicator{*;}
