--- app/src/main/java/com/vectras/vm/Blog/DataBlog.java
@@ -1,9 +1,0 @@
-package com.vectras.vm.Blog;
-
-public class DataBlog {
-
-    public String postThumb;
-    public String postTitle;
-    public String postContent;
-    public String postDate;
-}
