+++ app/src/main/java/com/vectras/vm/RomReceiverActivity.java
@@ -21,7 +21,7 @@ import java.util.Objects;
 import com.vectras.vm.utils.UIUtils;
 import com.vectras.vm.utils.PermissionUtils;
 
-public class ReceiveRomFileActivity extends AppCompatActivity {
+public class RomReceiverActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -38,7 +38,7 @@ public class ReceiveRomFileActivity extends AppCompatActivity {
                     startActivity(intent);
                     Toast.makeText(getApplicationContext(), getResources().getString(R.string.find_and_allow_access_to_storage_in_settings), Toast.LENGTH_LONG).show();
                 } else {
-                    ActivityCompat.requestPermissions(ReceiveRomFileActivity.this, new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE}, 1000);
+                    ActivityCompat.requestPermissions(RomReceiverActivity.this, new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE}, 1000);
                 }
             });
         }
@@ -69,7 +69,7 @@ public class ReceiveRomFileActivity extends AppCompatActivity {
                         Log.i("ReceiveRomFileActivity", uri.toString());
                         Log.i("ReceiveRomFileActivity", Objects.requireNonNull(getFilePath(uri)));
                     } else {
-                        Toast.makeText(ReceiveRomFileActivity.this, getResources().getString(R.string.format_not_supported_please_select_file_with_format_cvbi), Toast.LENGTH_LONG).show();
+                        Toast.makeText(RomReceiverActivity.this, getResources().getString(R.string.format_not_supported_please_select_file_with_format_cvbi), Toast.LENGTH_LONG).show();
                     }
                 } else if (Intent.ACTION_SEND.equals(action)) {
                     uri = intent.getParcelableExtra(Intent.EXTRA_STREAM);
@@ -88,12 +88,12 @@ public class ReceiveRomFileActivity extends AppCompatActivity {
                             Log.i("ReceiveRomFileActivity", uri.toString());
                             Log.i("ReceiveRomFileActivity", Objects.requireNonNull(getFilePath(uri)));
                         } else {
-                            Toast.makeText(ReceiveRomFileActivity.this, getResources().getString(R.string.format_not_supported_please_select_file_with_format_cvbi), Toast.LENGTH_LONG).show();
+                            Toast.makeText(RomReceiverActivity.this, getResources().getString(R.string.format_not_supported_please_select_file_with_format_cvbi), Toast.LENGTH_LONG).show();
                         }
                     }
                 }
             } else {
-                Toast.makeText(ReceiveRomFileActivity.this, getResources().getString(R.string.you_need_to_complete_vectras_vm_setup_before_importing_this_file), Toast.LENGTH_LONG).show();
+                Toast.makeText(RomReceiverActivity.this, getResources().getString(R.string.you_need_to_complete_vectras_vm_setup_before_importing_this_file), Toast.LENGTH_LONG).show();
                 startActivity(new Intent(this, SplashActivity.class));
             }
             finish();
