--- app/src/main/java/com/vectras/vm/Store/DataStore.java
@@ -1,13 +1,0 @@
-package com.vectras.vm.Store;
-
-public class DataStore {
-
-    public String itemName;
-    public String itemSize;
-    public String itemData;
-    public String itemIcon;
-    public String itemLink;
-    public String itemPreviewMain;
-    public String itemPreview1;
-    public String itemPreview2;
-}
