+++ app/src/main/java/com/vectras/vm/setupwizard/SetupWizard2Activity.java
@@ -33,8 +33,8 @@ import com.termux.app.TermuxService;
 import com.vectras.qemu.MainSettingsManager;
 import com.vectras.vm.AppConfig;
 import com.vectras.vm.R;
-import com.vectras.vm.RequestNetwork;
-import com.vectras.vm.RequestNetworkController;
+import com.vectras.vm.network.RequestNetwork;
+import com.vectras.vm.network.RequestNetworkController;
 import com.vectras.vm.databinding.ActivitySetupWizard2Binding;
 import com.vectras.vm.databinding.ListViewBinding;
 import com.vectras.vm.databinding.SetupQemuDoneBinding;
