+++ app/src/main/java/com/vectras/vm/setupwizard/SetupWizardActivity.java
@@ -35,8 +35,8 @@ import com.google.gson.reflect.TypeToken;
 import com.vectras.qemu.MainSettingsManager;
 import com.vectras.vm.AppConfig;
 import com.vectras.vm.R;
-import com.vectras.vm.RequestNetwork;
-import com.vectras.vm.RequestNetworkController;
+import com.vectras.vm.network.RequestNetwork;
+import com.vectras.vm.network.RequestNetworkController;
 import com.vectras.vm.databinding.ActivitySetupWizardBinding;
 import com.vectras.vm.databinding.SetupQemuAdvancedBinding;
 import com.vectras.vm.databinding.SetupQemuDoneBinding;
