--- app/src/main/res/layout/activity_profile.xml
@@ -1,105 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:id="@+id/main_layout"
-    android:fitsSystemWindows="true"
-    tools:context=".ProfileActivity">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-
-        <com.google.android.material.appbar.AppBarLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-
-            <androidx.appcompat.widget.Toolbar
-                android:id="@+id/toolbar"
-                android:layout_width="match_parent"
-                android:layout_height="?attr/actionBarSize" />
-
-            <ProgressBar
-                android:id="@+id/loadingPb"
-                style="?android:attr/progressBarStyleHorizontal"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:indeterminate="true"
-                android:visibility="gone" />
-        </com.google.android.material.appbar.AppBarLayout>
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="vertical">
-
-            <androidx.cardview.widget.CardView
-                android:layout_width="150dp"
-                android:layout_height="150dp"
-                android:layout_gravity="center_horizontal"
-                android:layout_margin="8dp"
-                android:clipToPadding="false"
-                style="@style/CardViewStyle"
-                android:focusable="true"
-                android:visibility="gone"
-                app:cardCornerRadius="360dp"
-                app:cardPreventCornerOverlap="true">
-
-                <ImageView
-                    android:id="@+id/profilePic"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:clickable="true"
-                    android:foreground="?android:attr/selectableItemBackground"
-                    android:scaleType="centerCrop"
-                    app:srcCompat="@drawable/person_24" />
-
-            </androidx.cardview.widget.CardView>
-
-
-            <com.google.android.material.textfield.TextInputLayout
-                android:id="@+id/profileNameField"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center_horizontal"
-                android:layout_marginBottom="4dp"
-                android:hint="@string/username"
-                android:theme="@style/AppTheme.TextInput"
-                app:errorContentDescription="username can't be empty!"
-                app:errorIconDrawable="@drawable/outline_bug_report_24">
-
-                <com.google.android.material.textfield.TextInputEditText
-                    android:id="@+id/profileName"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:imeActionLabel="@string/username"
-                    android:imeOptions="actionDone"
-                    android:inputType="textPersonName"
-                    android:singleLine="true"
-                    android:minWidth="150dp" />
-            </com.google.android.material.textfield.TextInputLayout>
-
-            <com.google.android.material.button.MaterialButton
-                android:id="@+id/saveBtn"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center_horizontal"
-                android:minWidth="150dp"
-                android:text="@string/save" />
-
-
-        </LinearLayout>
-    </LinearLayout>
-    <Button
-        android:id="@+id/signout"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/sign_out"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
