--- app/src/main/res/layout/container_post.xml
@@ -1,49 +1,0 @@
-<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/cdPost"
-	android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-	android:clickable="true"
-    android:foreground="?android:attr/selectableItemBackground"
-    style="@style/CardViewStyle"
-    app:cardCornerRadius="10dp"
-	app:cardUseCompatPadding="true"
-    android:focusable="true">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical" >
-
-        <ImageView
-            android:id="@+id/ivThumb"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:scaleType="centerInside" />
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:padding="10dp" >
-
-            <TextView
-                android:id="@+id/textTitle"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:maxLines="2"
-                android:text="Post title"
-                android:textSize="15sp"
-                android:textStyle="bold" />
-
-            <TextView
-                android:id="@+id/textDate"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="post date"
-                android:textSize="10sp" />
-        </LinearLayout>
-    </LinearLayout>
-
-</androidx.cardview.widget.CardView>
