+++ app/src/main/res/layout/create_vhd.xml
@@ -1,21 +1,41 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:padding="8dp">
+    android:background="?attr/colorSurfaceContainerLow"
+    android:padding="16dp">
+
+    <ImageView
+        android:id="@+id/icon"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:contentDescription="@string/create_a_new_virtual_disk"
+        android:src="@drawable/hard_drive_24px"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"/>
+
+    <TextView
+        android:id="@+id/title"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
+        android:textSize="20sp"
+        android:text="@string/create_a_new_virtual_disk"
+        app:layout_constraintStart_toEndOf="@id/icon"
+        app:layout_constraintTop_toTopOf="@id/icon"
+        app:layout_constraintBottom_toBottomOf="@id/icon"/>
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/nameField"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:hint="@string/image_name"
+        android:layout_marginTop="32dp"
+        android:hint="@string/file_name"
         android:theme="@style/AppTheme.TextInput"
-        app:startIconDrawable="@drawable/round_storage_24"
+        app:startIconDrawable="@drawable/edit_24px"
         app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
+        app:layout_constraintTop_toBottomOf="@id/title">
 
         <com.google.android.material.textfield.TextInputEditText
             android:id="@+id/name"
@@ -23,16 +43,18 @@
             android:layout_height="wrap_content"
             android:inputType="text"
             android:selectAllOnFocus="true"
-            android:singleLine="true" />
+            android:singleLine="true"
+            android:text="New virtual disk"/>
     </com.google.android.material.textfield.TextInputLayout>
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/sizeField"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:hint="@string/size_in_gb"
+        android:layout_marginTop="16dp"
+        android:hint="@string/size_gb"
         android:theme="@style/AppTheme.TextInput"
-        app:startIconDrawable="@drawable/round_storage_24"
+        app:startIconDrawable="@drawable/hard_drive_24px"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toBottomOf="@id/nameField">
@@ -43,23 +65,26 @@
             android:layout_height="wrap_content"
             android:inputType="number"
             android:selectAllOnFocus="true"
-            android:singleLine="true" />
+            android:singleLine="true"
+            android:text="128"/>
     </com.google.android.material.textfield.TextInputLayout>
-    <Button
-        android:id="@+id/createQcow2Btn"
-        style="@style/Widget.Material3Expressive.Button"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:enabled="false"
-        android:text="@string/create_qcow2"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/sizeField" />
+
     <TextView
         android:id="@+id/createPath"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:padding="8dp"
-        app:layout_constraintTop_toBottomOf="@id/createQcow2Btn"
+        android:layout_marginTop="16dp"
+        app:layout_constraintTop_toBottomOf="@id/sizeField"
         app:layout_constraintLeft_toLeftOf="parent"/>
+
+    <Button
+        android:id="@+id/createQcow2Btn"
+        style="@style/Widget.Material3Expressive.Button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="16dp"
+        android:text="@string/create"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/createPath" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
