--- app/src/main/res/layout/no_connection.xml
@@ -1,44 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/noConnectionLayout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:gravity="center_vertical|center|center_horizontal"
-    android:orientation="vertical"
-    android:visibility="gone">
-
-    <androidx.cardview.widget.CardView
-        style="@style/CardViewStyle"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:gravity="center_vertical|center|center_horizontal"
-        app:cardCornerRadius="10dp"
-        app:cardUseCompatPadding="true">
-
-        <LinearLayout
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:gravity="center_vertical|center|center_horizontal"
-            android:orientation="vertical"
-            android:padding="15dp">
-
-            <ImageView
-                android:layout_width="100dp"
-                android:layout_height="100dp"
-                android:layout_margin="5dp"
-                android:padding="15dp"
-                android:scaleType="fitXY"
-                android:src="@drawable/access_point_network_off"
-                app:srcCompat="@drawable/access_point_network_off" />
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:padding="5dp"
-                android:text="No Internet connection"
-                android:textSize="20sp"
-                android:textStyle="bold" />
-        </LinearLayout>
-    </androidx.cardview.widget.CardView>
-
-</LinearLayout>
