+++ app/src/main/res/menu/home_drawer_menu.xml
@@ -1,29 +1,4 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:id="@+id/navigation_item_store"
-        android:icon="@drawable/round_shop_24"
-        android:visible="false"
-        android:title="@string/store" />
-    <item
-        android:id="@+id/navigation_item_get_rom"
-        android:icon="@drawable/home_storage_24px"
-        android:visible="false"
-        android:title="@string/roms_store" />
-    <item
-        android:id="@+id/navigation_item_import_iso"
-        android:icon="@drawable/album_24px"
-        android:title="@string/import_iso"
-        android:visible="false"/>
-    <item
-        android:id="@+id/navigation_item_hdd1"
-        android:icon="@drawable/hard_drive_24px"
-        android:title="@string/hdd1"/>
-    <item
-        android:id="@+id/navigation_item_hdd2"
-        android:icon="@drawable/hard_drive_24px"
-        android:title="@string/hdd2"
-        android:visible="false"/>
     <item
         android:id="@+id/navigation_item_desktop"
         android:icon="@drawable/desktop_24px"
